Move overriden generator Word visitor
